Added Custom InputCache
For some reason, Gdx.input.isKeyJustPressed(keyCode) could return true multiple frames in a row. However, that functionality is needed to check for 1-time presses.

So, a custom cache was built which still uses Gdx.input under the hood, but properly handles isKeyJustPressed() to ensure it's never true 2 frames in a row (unless someone is toggling a key very quickly, at or faster than the framerate).